js: copyright script to gulp task
- convert copyright script to gulp task
- fix style problems
- move globby dependency to devDependencies
- add minimatch dependency to package.json
