+++ package.json
@@ -33,6 +33,7 @@
     "del": "^1.1.1",
     "duplexer": "^0.1.1",
     "event-stream": "^3.2.2",
+    "globby": "^1.2.0",
     "globule": "^0.2.0",
     "gulp": "^3.8.11",
     "gulp-clone": "^1.0.0",
@@ -61,6 +62,7 @@
     "main-bower-files": "^2.5.0",
     "map-stream": "^0.0.5",
     "marked": "^0.3.3",
+    "minimatch": "^2.0.4",
     "mkdirp": "^0.5.0",
     "mocha": "^2.1.0",
     "mocha-phantomjs": "^3.5.3",
@@ -94,7 +96,6 @@
     "csurf": "^1.7.0",
     "express": "^4.12.2",
     "express-session": "^1.10.3",
-    "globby": "^1.2.0",
     "ldapjs": "^0.7.1",
     "lodash": "~3.3.1",
     "lodash-deep": "^1.6.0",
